Created Enemies
Added enemies to the game. The player can attack and kill enemies and enemies are able to attack and kill the player. Enemy movement needs to be fixed so they aren't as fast. A fix for that can also be implemented for the player so the user can hold down the movement buttons instead of pressing it repeatedly.